Change application package
- Changed applicationId
- Changed the package from all application modules
