+++ base-android/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.urban.noise.base.android" />
+<manifest package="org.urban.noise.base.android" />
