+++ base-android/src/main/java/org/urban/noise/base/android/ui/base/BaseFragment.kt
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2019 urbannoise.org
+ *
+ * Licensed under the Attribution-NonCommercial-ShareAlike 4.0
+ * International (CC BY-NC-SA 4.0) you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://creativecommons.org/licenses/by-nc-sa/4.0/
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.urban.noise.base.android.ui.base
+
+import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+abstract class BaseFragment<B : ViewDataBinding, M : ViewModel>(
+    @LayoutRes
+    private val layoutId: Int
+) : Fragment() {
+
+    @Inject
+    lateinit var viewModel: M
+    lateinit var viewBinding: B
+
+    abstract fun onInitDependencyInjection()
+    abstract fun onInitDataBinding()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        viewBinding = DataBindingUtil.inflate(inflater, layoutId, container, false)
+        viewBinding.lifecycleOwner = viewLifecycleOwner
+        return viewBinding.root
+    }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        onInitDependencyInjection()
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        onInitDataBinding()
+    }
+
+    fun requireCompatActivity(): AppCompatActivity {
+        val activity = requireActivity()
+        if (activity is AppCompatActivity) {
+            return activity
+        } else {
+            throw TypeCastException("Main activity should extend from 'AppCompatActivity'")
+        }
+    }
+}
