+++ base-android/src/main/java/org/urban/noise/base/android/ui/bindings/ViewBindings.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2019 urbannoise.org
+ *
+ * Licensed under the Attribution-NonCommercial-ShareAlike 4.0
+ * International (CC BY-NC-SA 4.0) you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://creativecommons.org/licenses/by-nc-sa/4.0/
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.urban.noise.base.android.ui.bindings
+
+import android.view.View
+import android.view.View.GONE
+import android.view.View.INVISIBLE
+import android.view.View.VISIBLE
+import androidx.databinding.BindingAdapter
+
+@set:BindingAdapter("visible")
+var View.visible
+    get() = visibility == VISIBLE
+    set(value) {
+        visibility = if (value) VISIBLE else GONE
+    }
+
+@set:BindingAdapter("gone")
+var View.gone
+    get() = visibility == GONE
+    set(value) {
+        visibility = if (value) GONE else VISIBLE
+    }
+
+@set:BindingAdapter("invisible")
+var View.invisible
+    get() = visibility == INVISIBLE
+    set(value) {
+        visibility = if (value) INVISIBLE else VISIBLE
+    }
